Fix checkstyle violations in tests directory
- This should resolve checkstyle violations of all test files. Running
  './gradlew clean checkstyleTest' should show 0 violations.
