+++ src/test/java/randomTests/andre/segmentTree/test.java
@@ -3,15 +3,19 @@ package randomTests.andre.segmentTree;
 import dataStructures.segmentTree.SegmentTree;
 
 /**
- * Basic testing
+ * Basic testing.
  */
-
- public class test {
-    public static void main(String[] args) {
-        int[] t = new int[] {1,2,3,4,5,6,7};
-        SegmentTree tree = new SegmentTree(t);
-        System.out.println(tree.sumRange(2, 5));
-        tree.update(3, 10);
-        System.out.println(tree.sumRange(2, 5));
-    }
- }
+public class Test {
+  /**
+   * Custom test case for SegmentTree.
+   *
+   * @param args unused.
+   */
+  public static void main(String[] args) {
+    int[] t = new int[] {1, 2, 3, 4, 5, 6, 7};
+    SegmentTree tree = new SegmentTree(t);
+    System.out.println(tree.sumRange(2, 5));
+    tree.update(3, 10);
+    System.out.println(tree.sumRange(2, 5));
+  }
+}
